Übertragungsprotokoll ArduinoMicro <-> PX4 verbessert

+++ ArduCopter/UserVariables.h
@@ -13,8 +13,8 @@ int16_t iDistance = 0;
 
 int16_t iReadByte = 0;
 
-int8_t bStartByte = 255; 	// 1111 1111
-int8_t bStopByte = 254;		// 1111 1110
+int16_t bStartByte = 255; 	// 1111 1111
+int16_t bStopByte = 254;		// 1111 1110
 int8_t bValidFlags = 0;
 
 int16_t aiSensors[8] = {0};
